refactor(bedrock)!: move converse implementation into bedrock/converse
Relocate the root bedrock package (converse.go, converse_test.go,
modeloptions.go) into a new bedrock/converse subpackage and rename the
package from `bedrock` to `converse`, so the package name matches its
directory and leaves room for future bedrock/* subpackages.

- git mv the three files into bedrock/converse/.
- Rename package bedrock -> package converse.
- Update the OTel tracer name to the new import path.
- Update every example to import
  github.com/craigh33/adk-go-bedrock/bedrock/converse and call
  converse.New / converse.NewRuntimeAPI / converse.WithGuardrail, etc.
- Update README and example README references.

Breaking: consumers must import .../bedrock/converse and use the converse
package selector instead of bedrock.

Verified: go build ./..., go vet ./..., go test ./... -count=1,
golangci-lint run ./... (0 issues), and all examples compile.
