+++ examples/bedrock-a2a/main.go
@@ -26,7 +26,7 @@ import (
 
 	"go.opentelemetry.io/otel/trace"
 
-	"github.com/craigh33/adk-go-bedrock/bedrock"
+	"github.com/craigh33/adk-go-bedrock/bedrock/converse"
 	"github.com/craigh33/adk-go-bedrock/examples/internal/exampletrace"
 )
 
@@ -50,7 +50,7 @@ func newWeatherAgent(ctx context.Context, tp trace.TracerProvider) agent.Agent {
 	}
 
 	br := bedrockruntime.NewFromConfig(awsCfg)
-	llm, err := bedrock.NewWithAPI(modelID, bedrock.NewRuntimeAPI(br, bedrock.WithTracerProvider(tp)))
+	llm, err := converse.NewWithAPI(modelID, converse.NewRuntimeAPI(br, converse.WithTracerProvider(tp)))
 	if err != nil {
 		log.Fatalf("bedrock model: %v", err)
 	}
