+++ examples/bedrock-guardrails/main.go
@@ -19,7 +19,7 @@ import (
 	"google.golang.org/adk/v2/model"
 	"google.golang.org/genai"
 
-	"github.com/craigh33/adk-go-bedrock/bedrock"
+	"github.com/craigh33/adk-go-bedrock/bedrock/converse"
 	"github.com/craigh33/adk-go-bedrock/examples/internal/exampletrace"
 )
 
@@ -368,7 +368,7 @@ func main() {
 	defer func() { _ = shutdownTP(context.Background()) }()
 
 	br := bedrockruntime.NewFromConfig(awsCfg)
-	llm, err := bedrock.NewWithAPI(modelID, bedrock.NewRuntimeAPI(br, bedrock.WithTracerProvider(tp)))
+	llm, err := converse.NewWithAPI(modelID, converse.NewRuntimeAPI(br, converse.WithTracerProvider(tp)))
 	if err != nil {
 		log.Panicf("bedrock model: %v", err)
 	}
@@ -398,5 +398,5 @@ func main() {
 	fmt.Println("- FinishReason indicates if guardrails blocked the response")
 	fmt.Println("- Streaming accumulates guardrail metadata into final response")
 	fmt.Println("- Note: To use Bedrock guardrails, pre-configure a guardrail in AWS")
-	fmt.Println("  and pass its ID and version with bedrock.WithGuardrail")
+	fmt.Println("  and pass its ID and version with converse.WithGuardrail")
 }
