+++ examples/bedrock-video-gen/main.go
@@ -26,7 +26,7 @@ import (
 	"google.golang.org/adk/v2/tool"
 	"google.golang.org/genai"
 
-	"github.com/craigh33/adk-go-bedrock/bedrock"
+	"github.com/craigh33/adk-go-bedrock/bedrock/converse"
 	"github.com/craigh33/adk-go-bedrock/tools/videogenerator"
 )
 
@@ -63,7 +63,7 @@ func run(ctx context.Context) error {
 	br := bedrockruntime.NewFromConfig(awsCfg)
 	s3Client := s3.NewFromConfig(awsCfg)
 
-	llm, err := bedrock.NewWithAPI(modelID, bedrock.NewRuntimeAPI(br))
+	llm, err := converse.NewWithAPI(modelID, converse.NewRuntimeAPI(br))
 	if err != nil {
 		return fmt.Errorf("bedrock model: %w", err)
 	}
